requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.0
python-dotenv==1.0.0
requests==2.31.0
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.24.0
# bitsandbytes>=0.42.0
sentencepiece>=0.2.0
protobuf==4.24.4
# sounddevice==0.4.6
numpy>=1.26.2
scipy>=1.11.4
# librosa>=0.10.2
# SpeechRecognition==3.10.0
# pyaudio==0.2.14
tqdm==4.66.1
colorama==0.4.6
# Core dependencies your code uses
psycopg2-binary==2.9.10
bcrypt==4.3.0
PyJWT==2.10.1
groq==0.31.1
sentence-transformers==5.1.0
faiss-cpu==1.12.0
twilio==9.8.0
langdetect==1.0.9
python-multipart==0.0.20
